Carlo Bramini (carlo.bramix@libero.it) - Fix the display of Line and Column in the Status Bar and add language resources for this
Patch was slightly modified by me to also react on WM_KEYDOWN messages, so that the line/column is already updated, when you're holding down the buttons. (like Windows' notepad does by the way)
We still have to keep WM_KEYUP though, otherwise this doesn't work properly when being used with Page Up/Page Down.

See issue #3702 for more details.

svn path=/trunk/; revision=37176

+++ reactos/base/applications/notepad/settings.c
@@ -112,7 +112,7 @@ void LoadSettings(void)
 	base_length = (GetSystemMetrics(SM_CXSCREEN) > GetSystemMetrics(SM_CYSCREEN))?
 		GetSystemMetrics(SM_CYSCREEN) : GetSystemMetrics(SM_CXSCREEN);
 
-	dx = base_length * .95;
+	dx = (INT)(base_length * .95);
 	dy = dx * 3 / 4;
 	SetRect( &Globals.main_rect, 0, 0, dx, dy );
